Renamed devops_toolset module to avoid name conflicts
